DEF-56: make signal behavior consistent
